requirements.txt
Django
django-sslserver
python-decouple
django-import-export